Add `CreatedOnDateListFilter` to Series admin
Moved `CreatedOnDateListFilter` to utils to reuse code.
